requirements.txt
asgiref==3.4.1
astroid==2.8.4
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-admin-sortable2==1.0.2
django-tinymce==3.3.0
environs==9.3.4
idna==3.3
isort==5.9.3
lazy-object-proxy==1.6.0
marshmallow==3.14.0
mccabe==0.6.1
Pillow==8.4.0
platformdirs==2.4.0
pylint==2.11.1
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
wrapt==1.13.3
